+++ src/api/media/package.lisp
@@ -1,9 +1,4 @@
 (defpackage :CL-MATRIX.API.MEDIA (:use #:cl #:cl-matrix.autowrap.runtime #:cl-matrix.autowrap.authentication)
-  (:import-from :CL-MATRIX.AUTOWRAP.AUTHENTICATION
-  #:AUTH
-  #:ACCESS-TOKEN
-  #:HOMESERVER
-  #:QUERY-PARAM)
   (:export
   #:get-download/servername/mediaid
   #:get-thumbnail/servername/mediaid