+++ src/base-events.lisp
@@ -6,7 +6,7 @@
   (:nicknames "CL-M.BASE-EVENTS")
   (:export
    #:base-event
-   #:sync
+   #:sync-hook
    #:on-room
    #:room-event
    #:presence-event
@@ -20,9 +20,7 @@
 ))
 (in-package :cl-matrix.base-events)
 
-;;; why do we need to be passing account?
-;;; we at the moment bind *account* when we use the function issue-sync-event
-(defgeneric sync (account data))
+(defgeneric sync-hook (account data))
 
 (defgeneric on-room (account room-id room-data))
 
@@ -40,7 +38,7 @@
   (:method-combination standard)
   (:hook-point t))
 
-(method-hooks:defhook sync issue-on-room ((account account) data)
+(method-hooks:defhook sync-hook issue-on-room ((account account) data)
   (jsown:do-json-keys (room-id room-data) (jsown:filter data "rooms" "join")
     (on-room account room-id room-data)))
 
@@ -48,11 +46,11 @@
   (dolist (e (jsown:filter room-data "timeline" "events"))
     (room-event account room-id e)))
 
-(method-hooks:defhook sync issue-presence-events ((account account) data)
+(method-hooks:defhook sync-hook issue-presence-events ((account account) data)
   (dolist (e (jsown:filter data "presence" "events"))
     (presence-event account e)))
 
-(method-hooks:defhook sync issue-invite-events ((account account) data)
+(method-hooks:defhook sync-hook issue-invite-events ((account account) data)
   (let ((rooms (jsown:filter data "rooms" "invite")))
     (jsown:do-json-keys (room-id room-data) rooms
       (invite-event account room-id room-data))))
@@ -77,5 +75,5 @@
   (when (string= "m.text" (msgtype data))
     (m.text account room-id data)))
 
-(defun issue-sync-event (sync-data)
-  (sync *account* sync-data))
+(defun issue-sync-event (account sync-data)
+  (sync-hook account sync-data))